move helper functions into helper file
